COMP: Fix initial configuration on Linux
This commit is a follow of ff793c5 (COMP: Update Autoscoper to fix build
using gcc7 & specify OpenGL_GL_PREFERENCE) ensuring that OpenGL_GL_PREFERENCE
is initialized also when the variable is set to an empty string.

Since the "vtk-config" file saves & restores the value associated with
OpenGL_GL_PREFERENCE, the test implemented in "Autoscoper" external project
was failing to initialize the variable because it was defined and set to
an empty string.