BUG: Fix wrong transformations applied to generated PVs
At the end of the partial volume generation process, our approach for matching
the generated `*_t.tfm` and `*_scale.tfm` transforms to their corresponding
`*.tif` partial volume incorrectly relied on the order of the files as
found by `glob`.

The order of items returned by the function is arbitrary according to the
docs, so the index of each segment's corresponding volume and transforms in
the lists `vols`, `tfms_t` and `tfms_scale` were not guaranteed to be the
same. It sometimes resulted in the volume nodes being rendered with the wrong
transforms applied to them when volume generation was finished.

The code now specifically looks for the correct transforms corresponding to
the volume according to the node name.