ENH: Load combined scale and translation transforms to generated PVs
To reduce clutter for the app users, after the generation of partial
volumes in the pre-processing module, only the combined scale and
translation transformation layer is loaded and applied onto each PV,
instead of the two separate scale and translation trasnforms. I.e.,
`{}.tfm` is loaded and applied instead of `{}_t.ftm` and `{}_scale.tfm`.

All files are still written to the output directory as part of the PV
generation proceess as before.

Co-authored-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>