ENH: Support additional volume node as source for registration
The changes in this commit modify the registration workflow to register
the cropped region of an input source volume to each frame of the input
sequence. The previous wip implementation registered a cropped region
from the n-th frame in the sequence to its respective region in the
n+1-th frame.