BUG: Fix propagation of resulting elastix transform (again)
With the re-implementation of the workflow using a source volume, the
manually adjusted component of the initial guess transform wasn't
considered. This commit ensures that the total transform from the bone's
initial poisition in the frame to the target position (aka the
composition of the manual adjustment transform and the elastix output)
are propagated to the rest of the bones in the hierarchy.