BUG: Fix ignored transform when cropping volume from ROI
Prior to this change, cropping was done based on axes-aligned regions,
which means that the region to be cropped was determined using the min
and max coordinates of the transformed ROI, not based on the the whole
transformed geometry (see comment on GitHub PR). As a result, the
cropped target volume could potentially be larger than the cropped
source, sometimes leading to elastix failing with the error "Too many
samples map outside moving image buffer".