BUG: Reimplement import and export to be compatible with new workflow
As long as a TreeNode instance is saved in `rootNode` in the
`Hierarchical3DRegistrationWidget` class (aka a "session" is in
progress), the tracking results of the hierarchy can be exported. The
import functionality also works, but is a little more clunky. If an
existing session is in progress, importing will overwrite the current
transform sequence nodes in the scene with the imported data. If no
session is in progress, it will initialize one and load the transforms,
but the user won't be able to register over it. Future work can improve
the workflow to be more intentional with how users can import and resume
their past registration runs.